軽度なバグ修正

+++ HansoInputTool/Models/RowData.cs
@@ -7,8 +7,11 @@ namespace HansoInputTool.Models
         public int RowIndex { get; set; }
         public int? B_Day { get; set; }
         public int? C_Hanso { get; set; }
-        public double? D_YuryoKm { get; set; }
-        public double? E_MuryoKm { get; set; }
+
+        // double? から int? に変更
+        public int? D_YuryoKm { get; set; }
+        public int? E_MuryoKm { get; set; }
+
         public int? H_LateFeeOotsuki { get; set; }
         public int? K_LateMinutes { get; set; }
         public int? L_IsKoryo { get; set; }
